move string_extend_chars and clone_on_ref_ptr to their own module
